requirements.txt
praw
requests
feedparser
python-dotenv
openai
